+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: tmann <tmann@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/27 14:48:58 by tmann             #+#    #+#             */
-/*   Updated: 2019/03/16 18:46:21 by tmann            ###   ########.fr       */
+/*   Updated: 2019/03/18 19:43:01 by tmann            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -90,7 +90,7 @@ t_list				*ft_lstmap(t_list *lst, t_list *(*f) (t_list *elem));
 int					ft_countc(char *src, char c);
 char				*ft_l_itoa(long long n);
 char				*ft_ltoa_base(long long int nbr, int base);
-char				*ft_utoa_base(unsigned long long nbr, int base);
+char				*ft_utoa_base(unsigned long int nbr, int base);
 char				*ft_utoa_base_little(unsigned long long nbr, int base);
 
 #endif
